Subscribe and initialize state
The subscribe function allows parts of the application to register functions that should be called whenever the state changes. when the subscribe is called with a listener it add it to the  listeners array. 

The initialize state is dispatches an empty action this is to make sure that the reducer is called once to initialize the state to its default value.